requirements.txt
discord.py==2.3.0
motor==2.5.1
python-dotenv==0.21.0
aiohttp==3.8.1
pymongo==4.1.1
pytest==7.1.2
pytest-asyncio==0.18.3
prometheus_client==0.17.1
psutil==5.9.5
cachetools==5.3.1
colorlog==6.7.0
python-json-logger==2.0.7
aiodns==3.0.0  # For faster DNS resolution
cchardet==2.1.7  # For faster encoding detection
uvloop==0.17.0; sys_platform != "win32"  # Faster event loop for non-Windows systems
```